requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
defusedxml==0.7.1
Django==4.2.29
djangorestframework==3.16.1
idna==3.11
mysqlclient==2.2.8
python-dotenv==1.2.2
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
youtube-transcript-api==1.2.4
